requirements.txt
#note: no libraries for Windows available
amazon-braket-sdk
amazon-braket-algorithm-library
colorama
mpmath
numpy
setuptools
sympy